add basic unicode support
final result: basic symbols etc works. But not all langs work, either
due to font missing entries or combinations. A basic test shows
Russian working well, but Greek and Tamil having issues.

program anyhow spits out unicode based on the locale specification in
utf8
vterm needs to be told to operate in utf8 mode
finally, we get the cell data in UCS-4 format since we get 32 bits for
each character

the 6 entries is a bit of a confusion: it is not the max 6 8 bits
needed for utf-8 since each entry is 32 bits. It is instead one char +
6 combining characters

though we get the characters, I don't think that we can combine it
just using the xft lib without advanced shapers like harfbuzz. This,
though there is a XFTDrawStringUtf8 function, that does not do what we want